JOHNZON-226 register Adapters only once
Previous each MappingParserImpl has it's own set of ReverseAdapters.
They got rebuilt with every parsing (as this creates a new MappingParserImpl
each time.
I've now moved the ReverseAdapterMap to Config and populate it upfront.

I'm not yet fully happy with the solution as the ReverseAdapterMap also
contains dynamically registered Adapters. So feel free to move the Map
to a better place in the future.