JOHNZON-281 JAX-RS Provider MUST throw NoContentException
The Johnzon JAX-RS Provider throws the wrong exception type in case of reading empty input streams. JAX-RS 2.1 requires to throw NoContentException, but Johnzon throws JsonbException.

See JavaDocs of JAX-RS 2.1 MessageBodyReader:

"In case the entity input stream is empty, the reader is expected to either return a Java representation of a zero-length entity or throw a javax.ws.rs.core.NoContentException in case no zero-length entity representation is defined for the supported Java type."

Signed-off-by: Markus KARG <markus@headcrashing.eu>